source/backend/actuators_controller/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
docopt==0.6.2
fastapi==0.135.1
h11==0.16.0
idna==3.11
mariadb==1.1.14
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
stomp.py==8.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
websocket-client==1.9.0
requests
